chore: code cleanup phase 2 (REVIEW) — dedupe v1 ingest helpers
Both /v1/logs (OTLP) and /v1/ingest (simple) handlers carried two
byte-identical copies of the insert `.returning(...)` column map (31
columns, search tsvector excluded) and the same response-body builder.
Extract both into src/lib/server/utils/ingest.ts so the "source of
truth" is a single definition that cannot silently drift between the two
endpoints (a column added to one map only would otherwise diverge the
SSE payload shape for that endpoint).

Frozen queue:
- R1: add LOG_RETURNING_COLUMNS; both handlers call
  .returning(LOG_RETURNING_COLUMNS).
- R2: move buildIngestResponse into the shared util; /v1/ingest now uses
  it instead of an inline response build.

Behavior-preserving: column map and response shape are unchanged.

Validation (all match pre-cleanup baseline):
- bun run lint (vp check): 258 files, 0 warnings/errors
- bun run check (svelte-check): 0 errors, 0 warnings
- bun run knip: clean
- test:unit 355, test:component 352, test:integration 330 — all pass
