+++ src/routes/v1/ingest/+server.ts
@@ -7,6 +7,7 @@ import { log, project } from "$lib/server/db/schema";
 import { logEventBus, type StreamLog } from "$lib/server/events";
 import { ApiKeyError, validateApiKey } from "$lib/server/utils/api-key";
 import { requireJsonContentType } from "$lib/server/utils/content-type";
+import { buildIngestResponse, LOG_RETURNING_COLUMNS } from "$lib/server/utils/ingest";
 import {
   assignIncidentIds,
   prepareLogsForIncidents,
@@ -175,39 +176,7 @@ export const POST: RequestHandler = async ({ request, locals }) => {
           // eslint-disable-next-line @typescript-eslint/no-explicit-any
           const insertedLogs: StreamLog[] = await (
             tx.insert(log).values(logEntries) as any
-          ).returning({
-            id: log.id,
-            projectId: log.projectId,
-            incidentId: log.incidentId,
-            fingerprint: log.fingerprint,
-            serviceName: log.serviceName,
-            level: log.level,
-            message: log.message,
-            metadata: log.metadata,
-            timeUnixNano: log.timeUnixNano,
-            observedTimeUnixNano: log.observedTimeUnixNano,
-            severityNumber: log.severityNumber,
-            severityText: log.severityText,
-            body: log.body,
-            droppedAttributesCount: log.droppedAttributesCount,
-            flags: log.flags,
-            traceId: log.traceId,
-            spanId: log.spanId,
-            resourceAttributes: log.resourceAttributes,
-            resourceDroppedAttributesCount: log.resourceDroppedAttributesCount,
-            resourceSchemaUrl: log.resourceSchemaUrl,
-            scopeName: log.scopeName,
-            scopeVersion: log.scopeVersion,
-            scopeAttributes: log.scopeAttributes,
-            scopeDroppedAttributesCount: log.scopeDroppedAttributesCount,
-            scopeSchemaUrl: log.scopeSchemaUrl,
-            sourceFile: log.sourceFile,
-            lineNumber: log.lineNumber,
-            requestId: log.requestId,
-            userId: log.userId,
-            ipAddress: log.ipAddress,
-            timestamp: log.timestamp,
-          });
+          ).returning(LOG_RETURNING_COLUMNS);
           return { insertedLogs, touchedIncidents };
         })
       : { insertedLogs: [], touchedIncidents: [] };
@@ -220,12 +189,5 @@ export const POST: RequestHandler = async ({ request, locals }) => {
     logEventBus.emitIncident(touchedIncident);
   }
 
-  // Build response
-  const response: { accepted: number; rejected?: number; errors?: string[] } = { accepted };
-  if (rejected > 0) {
-    response.rejected = rejected;
-    response.errors = errors;
-  }
-
-  return json(response, { status: 200 });
+  return json(buildIngestResponse(accepted, rejected, errors), { status: 200 });
 };
